Remove python3.3 from appveyor testing
Python3.3 is currently not supported: In python3.3 handles are still
leaked (the solution will be to patch the rpc_util.file_monitor module
to be python3-compatible).

I will re-add this to testing if I find time to fix the problem.

Another reason for this change is that it decreases the (considerable)
appveyor build time by about a third.